Refactor config out from main_training and merge with upstream
